server/requirements.txt
# Web Framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Settings & Environment
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# Database
supabase>=2.4.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# LLM & AI
groq>=0.9.0
langchain>=0.2.0
langchain-groq>=0.1.6
langgraph>=0.1.0
google-generativeai>=0.7.0
sentence-transformers>=3.0.0

# HTTP Client
httpx>=0.27.0

# PDF Generation
weasyprint>=62.0
jinja2>=3.1.4

# Image Processing
Pillow>=10.3.0
cairosvg>=2.7.1

# Utilities
python-slugify>=8.0.4
zipfile36>=0.1.3
aiofiles>=23.2.1
